+++ footer.php
@@ -4,17 +4,22 @@
  *
  * @link https://developer.wordpress.org/reference/functions/get_footer/
  *
- * @package WordPress Development Environment ("WPDE")
+ * @package WordPress Development Environment (WPDE)
  * @author Jindřich Ručil
  * @since 1.0.0
  */
 ?>
 
 <?php
-if (!is_404() && !is_page('sign-in')) {
+if (!is_404()) {
+    get_template_part('template-parts/footer', 'top');
     get_template_part('template-parts/footer', 'main');
 }
-get_template_part('template-parts/modal', 'search');
+
+$search_form = get_field('search_form', 'option');
+if (empty($search_form)) {
+    get_template_part('template-parts/modal', 'search');
+}
 wp_footer();
 ?>
 </body>