+++ header.php
@@ -4,19 +4,12 @@
  *
  * @link https://developer.wordpress.org/reference/functions/get_header/
  *
- * @package WordPress Development Environment ("WPDE")
+ * @package WordPress Development Environment (WPDE)
  * @author Jindřich Ručil
  * @since 1.0.0
  */
 ?>
-<?php
-// Zobrazování všech chyb
-error_reporting(E_ALL);
 
-// Zajištění, že se chyby zobrazí na obrazovce
-ini_set('display_errors', 1);
-ini_set('display_startup_errors', 1);
-?>
 <!DOCTYPE html>
 <html <?php language_attributes(); ?> data-bs-theme="light" class="cc--lightmode">
 <head>