+++ search.php
@@ -4,7 +4,7 @@
  *
  * @link https://developer.wordpress.org/themes/basics/template-hierarchy/#search-result
  *
- * @package WordPress Development Environment ("WPDE")
+ * @package WordPress Development Environment (WPDE)
  * @author Jindřich Ručil
  * @since 1.0.0
  */
@@ -15,7 +15,7 @@ get_header();
 $search_query = get_search_query();
 ?>
 
-<div class="container-fluid px-0 py-5">
+<div class="container-fluid px-0 py-6">
     <div class="container">
         <div class="pb-5 mb-5 border-bottom">
             <small class="text-primary">
@@ -34,15 +34,16 @@ $search_query = get_search_query();
             <?php if (have_posts()) {
                 while (have_posts()) {
                     the_post();
-                    echo '<div class="col-md-4">';
-                    get_template_part('template-parts/content', 'post');
+                    $grid = get_field('grid_search') ? get_field('grid_search') : 4;
+                    echo '<div class="col-md-' . $grid . '">';
+                        get_template_part('template-parts/content', 'post');
                     echo '</div>';
                 }
                 wp_reset_postdata();
                 get_template_part('template-parts/content', 'pagination');
             } else {
                 $html = '<div class="col-lg-12">';
-                $html .= '<p class="text-danger mb-0">' . __('Sorry, no data was found matching your search terms. Please try again with different keywords.', 'wpde') . '</p>';
+                    $html .= '<p class="text-danger mb-0">' . __('Sorry, no data was found matching your search terms. Please try again with different keywords.', 'wpde') . '</p>';
                 $html .= '</div>';
 
                 echo $html;