+++ tag.php
@@ -4,7 +4,7 @@
  *
  * @link https://developer.wordpress.org/themes/basics/template-hierarchy/#category
  *
- * @package WordPress Development Environment ("WPDE")
+ * @package WordPress Development Environment (WPDE)
  * @author Jindřich Ručil
  * @since 1.0.0
  */
@@ -12,7 +12,7 @@
 
 <?php get_header(); ?>
 
-<div class="container-fluid px-0 py-5">
+<div class="container-fluid px-0 py-6">
     <div class="container">
         <div class="pb-5 mb-4 border-bottom">
             <small class="text-primary">
@@ -32,8 +32,9 @@
             if (have_posts()) {
                 while (have_posts()) {
                     the_post();
-                    echo '<div class="col-md-4">';
-                    get_template_part('template-parts/content', 'post');
+                    $grid = get_field('grid_tag') ? get_field('grid_tag') : 4;
+                    echo '<div class="col-md-' . $grid . '">';
+                        get_template_part('template-parts/content', 'post');
                     echo '</div>';
                 }
                 wp_reset_postdata();