+++ template-parts/footer-main.php
@@ -3,7 +3,10 @@
 		<div class="container">
 			<div class="d-flex flex-column flex-sm-row justify-content-start justify-content-md-between align-items-center py-4 mt-4 border-top">
 				<div>
-				<small class="text-muted">© <?php echo date('Y'); ?> <?php bloginfo('name'); ?>. <?php _e('All rights reserved.', 'wpde'); ?></small>
+					<?php
+					$copyright = get_field('logo_text', 'option') ? get_field('logo_text', 'option') : bloginfo('name');
+					?>
+				<small class="text-muted">© <?php echo date('Y') . ' ' . $copyright . '. ' . __('All rights reserved.', 'wpde'); ?></small>
 				</div>
 				<div class="d-flex align-items-center gap-4">
 					<ul class="list-unstyled d-flex align-items-center flex-wrap flex-md-nowrap column-gap-3 mb-0">