+++ changelog.md
@@ -1,3 +1,11 @@
+## Version 1.0.1 (Fri, July 19 2024)
+
+- NEW! - Settings UI
+- NEW! - Settings fields
+- NEW! - Grid system
+- NEW! - Gallery section
+- FIX! - 88+ bugs fixed
+
 ## Version 1.0.0 (Fri, May 24 2024)
 
--   Initial public release
+- Initial public release